test: 게임 방 초기화 테스트를 Room 도메인으로 이동
